kinda fixed everything

+++ video_gfx/src/helpers/remove_folder.py
@@ -1,11 +1,11 @@
+import shutil
 from pathlib import Path
 
 
 def remove_tree(path: Path) -> None:
-    for item in path.glob("*"):
-        if item.is_file():
-            item.unlink()
-        elif item.is_dir():
-            remove_tree(item)
-            item.rmdir()
-    # path.rmdir()
+    if path.exists():
+        for item in path.glob("*"):
+            if item.is_file():
+                item.unlink()
+            elif item.is_dir():
+                shutil.rmtree(item)
